+++ examples/src/main/scala/org/apache/spark/examples/sql/streaming/StructuredNetworkWordCount.scala
@@ -25,12 +25,13 @@ import org.apache.spark.sql.SparkSession
  * Counts words in UTF8 encoded, '\n' delimited text received from the network every second.
  *
  * Usage: StructuredNetworkWordCount <hostname> <port> <checkpoint dir>
- * <hostname> and <port> describe the TCP server that Spark Streaming would connect to receive data.
+ * <hostname> and <port> describe the TCP server that Structured Streaming
+ * would connect to receive data.
  *
  * To run this on your local machine, you need to first run a Netcat server
  *    `$ nc -lk 9999`
  * and then run the example
- *    `$ bin/run-example org.apache.spark.examples.sql.streaming.StructuredNetworkWordCount
+ *    `$ bin/run-example sql.streaming.StructuredNetworkWordCount
  *    localhost 9999 <checkpoint dir>`
  */
 object StructuredNetworkWordCount {
